Development into main  (#1400)
* Qt6 migration (#1375)

* Qt6 migration

* download, build and install qt6

* removed no longer supported command line argument

* updated the cmake version to 3.24.4

* not static

* xcb lib

* missing xcb lib

* Update install_ubuntu_dependencies_build.sh

* Update main.yml to cat the valgrind_gui.log

* QVariant deprecated member

* Fixed migration issues

* Replace qAsConst with std::as_const

* Fix centos qt6 (#1364)

Fixed CentOS build Qt6

---------


* Fixed build issues

* Incremented patch version

* Fixed valgrind build

* Replace hasPendingEvents()

* Incremented patch version

---------


* Rework task binding to tcl command to make it more readable

* Incremented patch version

* Remove copy/paste code from ProjectManager::setSimulationFiles, ProjectManager::setDesignFiles

* Incremented patch version

* Fixed build after merge

* Rename Detailed to Placement for compiler action enum value

* Incremented patch version

* Remove copy/paste code from IpConfigWidget

* return TIMING_INFO extraction for task message. this part was presented in development(qt5) branch

* sequential_map::push_back removes previosly added values with same key, otherwise the methods sequential_map::operator[] and sequential_map::value() will return value for the first(oldest) occured key

* sequential_map.h fix formatting with clang-format

* Incremented patch version

* Add sequential_multi_map

* Fixed ClangTidy CI action

* Incremented patch version

* [RG-422] Add clean all action (#1401)

* Macos qt6 issue (#1402)

---------

+++ src/Utils/FileUtils.cpp
@@ -397,7 +397,10 @@ bool FileUtils::removeFile(const std::filesystem::path& file) noexcept {
 
 bool FileUtils::removeAll(const std::filesystem::path& path) {
   bool ok{true};
-  for (auto const& dir_entry : std::filesystem::directory_iterator{path}) {
+  std::error_code ec{};
+  std::filesystem::directory_iterator dir{path, ec};
+  if (ec) return false;
+  for (auto const& dir_entry : dir) {
     ok &= FileUtils::removeFile(dir_entry);
   }
   return ok;
